+++ Growy.Function/Models/Dtos/DevelopmentReportRequest.cs
@@ -4,8 +4,8 @@ public record DevelopmentReportRequest
 {
     public Guid ParentId { get; set; }
     public Guid ChildId { get; set; }
-    public List<int> Answers { get; init; } = new();
-    public int TotalMentalAge { get; init; }
+    public List<int> Answers { get; init; } = new(); 
+    public float TotalMentalAge { get; init; }
     public float DqResult { get; init; }
     public float CandidateMonth { get; init; }
 }
